+++ server/lib/prosody/config.ts
@@ -66,6 +66,7 @@ async function getProsodyFilePaths (options: RegisterServerOptions): Promise<Pro
 interface ProsodyConfig {
   content: string
   paths: ProsodyFilePaths
+  host: string
   port: string
   baseApiUrl: string
 }
@@ -77,6 +78,7 @@ async function getProsodyConfig (options: RegisterServerOptions): Promise<Prosod
   if (!/^\d+$/.test(port)) {
     throw new Error('Invalid port')
   }
+  const enableC2s = (await options.settingsManager.getSetting('prosody-c2s') as boolean) || false
   const prosodyDomain = await getProsodyDomain(options)
   const paths = await getProsodyFilePaths(options)
 
@@ -99,10 +101,20 @@ async function getProsodyConfig (options: RegisterServerOptions): Promise<Prosod
   config.usePeertubeBosh(prosodyDomain, port)
   config.useMucHttpDefault(roomApiUrl)
 
+  if (enableC2s) {
+    const c2sPort = (await options.settingsManager.getSetting('prosody-c2s-port') as string) || '52822'
+    if (!/^\d+$/.test(c2sPort)) {
+      throw new Error('Invalid c2s port')
+    }
+    config.useC2S(c2sPort)
+  }
+
   // TODO: add a settings so that admin can choose? (on/off and duration)
   config.useMam('1w') // Remove archived messages after 1 week
   // TODO: add a settings to choose?
   config.useDefaultPersistent()
+
+  config.useListRoomsApi(apikey)
 
   config.useTestModule(apikey, testApiUrl)
 
@@ -126,7 +138,8 @@ async function getProsodyConfig (options: RegisterServerOptions): Promise<Prosod
     content,
     paths,
     port,
-    baseApiUrl
+    baseApiUrl,
+    host: prosodyDomain
   }
 }