+++ server/lib/routers/webchat.ts
@@ -1,16 +1,23 @@
 import type { Router, RequestHandler, Request, Response, NextFunction } from 'express'
 import type { ProxyOptions } from 'express-http-proxy'
-import type { ChatType } from '../../../shared/lib/types'
-import { getBaseRouterRoute, getBaseStaticRoute } from '../helpers'
+import type { ChatType, ProsodyListRoomsResult } from '../../../shared/lib/types'
+import { getBaseRouterRoute, getBaseStaticRoute, isUserAdmin } from '../helpers'
 import { asyncMiddleware } from '../middlewares/async'
 import { getProsodyDomain } from '../prosody/config/domain'
+import { getAPIKey } from '../apikey'
 import * as path from 'path'
 const bodyParser = require('body-parser')
+const got = require('got')
 
 const fs = require('fs').promises
 const proxy = require('express-http-proxy')
 
 let httpBindRoute: RequestHandler
+interface ProsodyHttpBindInfo {
+  host: string
+  port: string
+}
+let currentProsodyHttpBindInfo: ProsodyHttpBindInfo | null = null
 
 async function initWebchatRouter (options: RegisterServerOptions): Promise<Router> {
   const {
@@ -97,22 +104,76 @@ async function initWebchatRouter (options: RegisterServerOptions): Promise<Route
       httpBindRoute(req, res, next)
     }
   )
+
+  router.get('/prosody-list-rooms', asyncMiddleware(
+    async (req: Request, res: Response, _next: NextFunction) => {
+      if (!res.locals.authenticated) {
+        res.sendStatus(403)
+        return
+      }
+      if (!await isUserAdmin(options, res)) {
+        res.sendStatus(403)
+        return
+      }
+
+      const chatType: ChatType = await options.settingsManager.getSetting('chat-type') as ChatType
+      if (chatType !== 'builtin-prosody') {
+        const message = 'Please save the settings first.' // TODO: translate?
+        res.status(200)
+        const r: ProsodyListRoomsResult = {
+          ok: false,
+          error: message
+        }
+        res.json(r)
+        return
+      }
+
+      if (!currentProsodyHttpBindInfo) {
+        throw new Error('It seems that prosody is not binded... Cant list rooms.')
+      }
+      const apiUrl = `http://localhost:${currentProsodyHttpBindInfo.port}/peertubelivechat_list_rooms/list-rooms`
+      peertubeHelpers.logger.debug('Calling list rooms API on url: ' + apiUrl)
+      const rooms = await got(apiUrl, {
+        method: 'GET',
+        headers: {
+          authorization: 'Bearer ' + await getAPIKey(options),
+          host: currentProsodyHttpBindInfo.host
+        },
+        responseType: 'json',
+        resolveBodyOnly: true
+      })
+
+      res.status(200)
+      const r: ProsodyListRoomsResult = {
+        ok: true,
+        rooms: rooms
+      }
+      res.json(r)
+    }
+  ))
+
   return router
 }
 
-function changeHttpBindRoute ({ peertubeHelpers }: RegisterServerOptions, port: string | null): void {
+function changeHttpBindRoute (
+  { peertubeHelpers }: RegisterServerOptions,
+  prosodyHttpBindInfo: ProsodyHttpBindInfo | null
+): void {
   const logger = peertubeHelpers.logger
-  logger.info('Changing http-bind port for ' + (port ?? 'null'))
-  if (port !== null && !/^\d+$/.test(port)) {
-    logger.error('Port is not valid. Replacing by null')
-    port = null
+  if (prosodyHttpBindInfo && !/^\d+$/.test(prosodyHttpBindInfo.port)) {
+    logger.error(`Port '${prosodyHttpBindInfo.port}' is not valid. Replacing by null`)
+    prosodyHttpBindInfo = null
   }
-  if (port === null) {
+
+  if (!prosodyHttpBindInfo) {
+    logger.info('Changing http-bind port for null')
+    currentProsodyHttpBindInfo = null
     httpBindRoute = (_req: Request, res: Response, _next: NextFunction) => {
       res.status(404)
       res.send('Not found')
     }
   } else {
+    logger.info('Changing http-bind port for ' + prosodyHttpBindInfo.port + ', on host ' + prosodyHttpBindInfo.host)
     const options: ProxyOptions = {
       https: false,
       proxyReqPathResolver: async (_req: Request): Promise<string> => {
@@ -122,7 +183,8 @@ function changeHttpBindRoute ({ peertubeHelpers }: RegisterServerOptions, port: 
       parseReqBody: true // Note that setting this to false overrides reqAsBuffer and reqBodyEncoding below.
       // FIXME: should we remove cookies?
     }
-    httpBindRoute = proxy('http://localhost:' + port, options)
+    currentProsodyHttpBindInfo = prosodyHttpBindInfo
+    httpBindRoute = proxy('http://localhost:' + prosodyHttpBindInfo.port, options)
   }
 }