+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v3.2.0
+
+### Features
+
+* Builtin Prosody: list existing rooms in the settings page
+* Builtin Prosody: new settings to enable local C2S. For example, can be used with Matterbridge (thanks https://github.com/tytan652)
+
+### Fixes
+
+* Fix broken API diagnostic.
+
 ## v3.1.0
 
 ### Features